+++ software/3D-CMCC-Forest-Model/src/canopy_evapotranspiration.c
@@ -14,7 +14,7 @@
 #include "constants.h"
 
 
-extern void Get_canopy_evapotranspiration (SPECIES *const s,  CELL *const c, const MET_DATA *const met, int month, int day, int DaysInMonth, float vpd, int height, int age, int species)
+extern void Get_canopy_evapotranspiration (SPECIES *const s,  CELL *const c, const MET_DATA *const met, int month, int day, int DaysInMonth, double vpd, int height, int age, int species)
 {
 
 	static int i;
@@ -30,9 +30,9 @@ extern void Get_canopy_evapotranspiration (SPECIES *const s,  CELL *const c, con
 			if (c->heights[height].z == c->top_layer)
 			{
 				s->value[CANOPY_EVAPOTRANSPIRATION] = s->value[DAILY_TRANSP] + s->value[RAIN_INTERCEPTED];
-				Log("Canopy evapotranspiration = %g mm\n", s->value[CANOPY_EVAPOTRANSPIRATION]);
+				Log("Canopy evapotranspiration = %f mm\n", s->value[CANOPY_EVAPOTRANSPIRATION]);
 				c->daily_c_evapotransp[c->top_layer] += s->value[CANOPY_EVAPOTRANSPIRATION];
-				Log("Canopy evapotranspiration from dominant layer = %g mm \n", c->daily_c_evapotransp[c->top_layer]);
+				Log("Canopy evapotranspiration from dominant layer = %f mm \n", c->daily_c_evapotransp[c->top_layer]);
 				/*last height dominant class processed*/
 				if (c->dominant_veg_counter == c->height_class_in_layer_dominant_counter)
 				{
@@ -51,9 +51,9 @@ extern void Get_canopy_evapotranspiration (SPECIES *const s,  CELL *const c, con
 				if (c->heights[height].z == c->top_layer-1)
 				{
 					s->value[CANOPY_EVAPOTRANSPIRATION] = s->value[DAILY_TRANSP] + s->value[RAIN_INTERCEPTED];
-					Log("Canopy evapotranspiration = %g mm\n", s->value[CANOPY_EVAPOTRANSPIRATION]);
+					Log("Canopy evapotranspiration = %f mm\n", s->value[CANOPY_EVAPOTRANSPIRATION]);
 					c->daily_c_evapotransp[c->top_layer-1] += s->value[CANOPY_EVAPOTRANSPIRATION];
-					Log("Canopy evapotranspiration from dominant layer = %g mm \n", c->daily_c_evapotransp[c->top_layer-1]);
+					Log("Canopy evapotranspiration from dominant layer = %f mm \n", c->daily_c_evapotransp[c->top_layer-1]);
 					/*last height dominant class processed*/
 					if (c->dominated_veg_counter == c->height_class_in_layer_dominated_counter)
 					{
@@ -69,9 +69,9 @@ extern void Get_canopy_evapotranspiration (SPECIES *const s,  CELL *const c, con
 				else
 				{
 					s->value[CANOPY_EVAPOTRANSPIRATION] = s->value[DAILY_TRANSP] + s->value[RAIN_INTERCEPTED];
-					Log("Canopy evapotranspiration = %g mm\n", s->value[CANOPY_EVAPOTRANSPIRATION]);
+					Log("Canopy evapotranspiration = %f mm\n", s->value[CANOPY_EVAPOTRANSPIRATION]);
 					c->daily_c_evapotransp[c->top_layer-2] += s->value[CANOPY_EVAPOTRANSPIRATION];
-					Log("Canopy evapotranspiration from dominant layer = %g mm \n", c->daily_c_evapotransp[c->top_layer-2]);
+					Log("Canopy evapotranspiration from dominant layer = %f mm \n", c->daily_c_evapotransp[c->top_layer-2]);
 					/*last height dominant class processed*/
 					if (c->subdominated_veg_counter == c->height_class_in_layer_subdominated_counter)
 					{
@@ -94,14 +94,14 @@ extern void Get_canopy_evapotranspiration (SPECIES *const s,  CELL *const c, con
 	else
 	{
 		s->value[CANOPY_EVAPOTRANSPIRATION] = 0.0;
-		Log("Canopy evapotranspiration = %g mm\n", s->value[CANOPY_EVAPOTRANSPIRATION]);
+		Log("Canopy evapotranspiration = %f mm\n", s->value[CANOPY_EVAPOTRANSPIRATION]);
 		c->daily_c_evapotransp[c->heights[height].z] = 0.0;
-		Log("Evapotraspirated water from layer %d = %g mm \n", c->heights[height].z, c->daily_c_evapotransp[c->heights[height].z]);
+		Log("Evapotraspirated water from layer %d = %f mm \n", c->heights[height].z, c->daily_c_evapotransp[c->heights[height].z]);
 	}
 
 	/*compute total daily canopy evapotranspiration*/
 	c->daily_tot_c_evapotransp += c->daily_c_evapotransp[c->heights[height].z];
-	Log("Daily total canopy evapotranspiration = %g \n", c->daily_tot_c_evapotransp);
+	Log("Daily total canopy evapotranspiration = %f \n", c->daily_tot_c_evapotransp);
 
 	i = c->heights[height].z;
